Work on background tasks for saving results (WIP, inefficient & buggy).
Also split SearchContext & app search strategies into separate files.
