+++ app/api/skaupat/query.py
@@ -88,50 +88,3 @@ def build_product_search_vars(
         "slugs": query["category"],
         "limit": limit
     }
-
-
-async def send_product_query(
-        query: dict[str, str], params: dict[str, Any]
-        ) -> tuple[
-            dict[str, str | int],
-            list[
-                tuple[
-                    schemas.Product,
-                    schemas.ProductData
-                ]
-            ]
-        ]:
-    """Send a product query to the external API and then parse it.
-
-    This function is essentially a wrapper for the request.send_request() func.
-    The intent is to allow the call to asyncio.gather() to parse the retrieved
-    responses immediately, while still waiting for the others to finish.
-
-    TODO: Consider a better location to place this function.
-
-
-    Args:
-        query (dict[str, str]):
-        A dict containing the query string & query category.
-        params (dict[str, Any]):
-        The request & query parameters to be used with the request.
-
-    Returns:
-        tuple[
-            dict[str, str | int],
-            list[
-                tuple[
-                    schemas.Product,
-                    schemas.ProductData
-                ]
-            ]
-        ]:
-        Return a tuple with the first item being a dict with the query details.
-        The second item is a list with all the parsed product items inside it.
-        Each parsed item is a tuple containing two different pydantic schemas.
-    """
-    response = await request.send_request(params=params)
-    return parse.parse_product_response(
-        response=response,
-        query=query
-    )
